WIP wasm interface

+++ p2panda-rs/src/wasm.rs
@@ -290,17 +290,14 @@ struct SignEncodeEntryResult {
 ///
 /// `entry_backlink_hash`, `entry_skiplink_hash`, `seq_num` and `log_id` are obtained by querying
 /// the `getEntryArguments` method of a p2panda node.
-///
-/// `seq_num` and `log_id` are `i32` parameters even though they have 64 bits in the Bamboo spec.
-/// Webkit doesn't support `BigInt` so it can't handle those large values.
 #[wasm_bindgen(js_name = signEncodeEntry)]
 pub fn sign_encode_entry(
     key_pair: &KeyPair,
     encoded_operation: String,
     entry_skiplink_hash: Option<String>,
     entry_backlink_hash: Option<String>,
-    seq_num: i32,
-    log_id: i32,
+    seq_num: u64,
+    log_id: u64,
 ) -> Result<JsValue, JsValue> {
     // If skiplink_hash exists construct Hash
     let skiplink_hash = match entry_skiplink_hash {
